More refactoring...
