+++ packages/manager/src/features/Domains/DomainDetail/DomainRecords/DomainRecords.styles.ts
@@ -1,12 +1,10 @@
 import { styled } from '@mui/material/styles';
 import Grid from '@mui/material/Unstable_Grid2';
+
 import { TableCell } from 'src/components/TableCell';
 
 export const StyledGrid = styled(Grid, { label: 'StyledGrid' })(
   ({ theme }) => ({
-    margin: 0,
-    marginTop: theme.spacing(2),
-    width: '100%',
     '& .MuiGrid-item': {
       paddingLeft: 0,
       paddingRight: 0,
@@ -16,30 +14,33 @@ export const StyledGrid = styled(Grid, { label: 'StyledGrid' })(
         marginRight: theme.spacing(),
       },
     },
+    margin: 0,
+    marginTop: theme.spacing(2),
     [theme.breakpoints.down('md')]: {
       marginLeft: theme.spacing(),
       marginRight: theme.spacing(),
     },
+    width: '100%',
   })
 );
 
 export const StyledTableCell = styled(TableCell, { label: 'StyledTabelCell' })(
   ({ theme }) => ({
-    whiteSpace: 'nowrap' as const,
-    width: 'auto',
     '& .data': {
       maxWidth: 300,
       overflow: 'hidden',
       textOverflow: 'ellipsis',
-      whiteSpace: 'nowrap' as const,
       [theme.breakpoints.up('md')]: {
         maxWidth: 750,
       },
+      whiteSpace: 'nowrap' as const,
     },
     '&:last-of-type': {
       display: 'flex',
       justifyContent: 'flex-end',
     },
+    whiteSpace: 'nowrap' as const,
+    width: 'auto',
   })
 );
